custom-elements/registering-custom-elements/definition-construction-algorithm-test-007.html

<!DOCTYPE html>
<html>
<head>
<title>If NAME is null then localName must be set to TYPE</title>
<meta name="author" title="Vasiliy Degtyarev" href="mailto:vasya@unipro.ru">
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If NAME is null then localName must be set to TYPE">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();    
    
    for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
        var name = 'x-1-' + HTML5_ELEMENTS[i];
        
        var element = doc.createElement(HTML5_ELEMENTS[i]);
        element.removeAttribute('localName');
        
        var proto = Object.create(element.constructor.prototype);
        
        var GeneratedConstructor = doc.register(name, {prototype: proto});
        var customElement = new GeneratedConstructor();
        
        assert_equals(customElement.localName, name, 'LocalName should be a type in case of ' + 
                'attempt to register a custom element if name is null');
    }    
    
}, 'If NAME is not specified then localName must be set to TYPE');


test(function() {
    var doc = newHTMLDocument();    
    
    for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
        var name = 'x-2-' + HTML5_ELEMENTS[i];
        
        var element = doc.createElement(HTML5_ELEMENTS[i]);
        element.removeAttribute('localName');
        
        var proto = Object.create(element.constructor.prototype);
        
        var GeneratedConstructor = doc.register(name, {prototype: proto, extends: null});
        var customElement = new GeneratedConstructor();
        
        assert_equals(customElement.localName, name, 'LocalName should be a type in case of ' + 
                'attempt to register a custom element if name is null');
    }    
    
}, 'If NAME is null then localName must be set to TYPE');

</script>
</body>
</html>
